requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.104.0
python-docx==1.0.1
PyPDF2==4.0.1
python-dotenv==1.0.0
openai==1.3.0
spacy==3.7.2
requests==2.31.0
pandas==2.1.3
